+++ crypto/CMakeLists.txt
@@ -348,7 +348,6 @@ add_library(
   evp_extra/p_ec_asn1.c
   evp_extra/p_ed25519.c
   evp_extra/p_ed25519_asn1.c
-  evp_extra/p_hkdf.c
   evp_extra/p_rsa_asn1.c
   evp_extra/p_x25519.c
   evp_extra/p_x25519_asn1.c
