+++ crypto/evp_extra/p_methods.c
@@ -10,7 +10,6 @@
 static const EVP_PKEY_METHOD *const non_fips_pkey_evp_methods[] = {
   &ed25519_pkey_meth,
   &x25519_pkey_meth,
-  &hkdf_pkey_meth,
 };
 
 static const EVP_PKEY_ASN1_METHOD *const asn1_evp_pkey_methods[] = {
