+++ crypto/fipsmodule/bcm.c
@@ -103,6 +103,7 @@
 #include "evp/evp.c"
 #include "evp/evp_ctx.c"
 #include "evp/p_ec.c"
+#include "evp/p_hkdf.c"
 #include "evp/p_rsa.c"
 #include "hkdf/hkdf.c"
 #include "hmac/hmac.c"
