+++ crypto/fipsmodule/evp/internal.h
@@ -246,8 +246,8 @@ struct evp_pkey_method_st {
   int (*ctrl)(EVP_PKEY_CTX *ctx, int type, int p1, void *p2);
 } /* EVP_PKEY_METHOD */;
 
-#define FIPS_EVP_PKEY_METHODS 3
-#define NON_FIPS_EVP_PKEY_METHODS 3
+#define FIPS_EVP_PKEY_METHODS 4
+#define NON_FIPS_EVP_PKEY_METHODS 2
 #define ASN1_EVP_PKEY_METHODS 6
 
 struct fips_evp_pkey_methods {
@@ -257,6 +257,7 @@ struct fips_evp_pkey_methods {
 const EVP_PKEY_METHOD *EVP_PKEY_rsa_pkey_meth(void);
 const EVP_PKEY_METHOD *EVP_PKEY_rsa_pss_pkey_meth(void);
 const EVP_PKEY_METHOD *EVP_PKEY_ec_pkey_meth(void);
+const EVP_PKEY_METHOD *EVP_PKEY_hkdf_pkey_meth(void);
 
 #if defined(__cplusplus)
 }  // extern C
