+++ crypto/fipsmodule/rand/urandom_test.cc
@@ -18,6 +18,7 @@
 #include <openssl/rand.h>
 
 #include "internal.h"
+#include "../../evp_extra/internal.h"
 #include "getrandom_fillin.h"
 
 #if defined(OPENSSL_X86_64) && !defined(BORINGSSL_SHARED_LIBRARY) && \
